+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-checkbox/dynamic-form-checkbox.component.ts
@@ -1,4 +1,3 @@
-import { NgIf } from '@angular/common';
 import { Component } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormCheckbox, DynamicFormInputBase, DynamicFormValidationService } from '@dynamic-forms/core';
@@ -7,7 +6,7 @@ import { DynamicFormCheckbox, DynamicFormInputBase, DynamicFormValidationService
   standalone: true,
   selector: 'bs-dynamic-form-checkbox',
   templateUrl: './dynamic-form-checkbox.component.html',
-  imports: [NgIf, ReactiveFormsModule],
+  imports: [ReactiveFormsModule],
 })
 export class BsDynamicFormCheckboxComponent extends DynamicFormInputBase<DynamicFormCheckbox> {
   constructor(protected override validationService: DynamicFormValidationService) {