+++ libs/bootstrap/src/lib/dynamic-form-input/dynamic-form-radio/dynamic-form-radio.component.ts
@@ -1,4 +1,3 @@
-import { NgFor } from '@angular/common';
 import { Component } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { DynamicFormInputBase, DynamicFormRadio, DynamicFormValidationService } from '@dynamic-forms/core';
@@ -7,7 +6,7 @@ import { DynamicFormInputBase, DynamicFormRadio, DynamicFormValidationService } 
   standalone: true,
   selector: 'bs-dynamic-form-radio',
   templateUrl: './dynamic-form-radio.component.html',
-  imports: [NgFor, ReactiveFormsModule],
+  imports: [ReactiveFormsModule],
 })
 export class BsDynamicFormRadioComponent extends DynamicFormInputBase<DynamicFormRadio> {
   constructor(protected override validationService: DynamicFormValidationService) {