+++ libs/material/src/lib/dynamic-form-action/dynamic-form-icon/dynamic-form-icon.component.ts
@@ -1,4 +1,4 @@
-import { NgClass, NgIf } from '@angular/common';
+import { NgClass } from '@angular/common';
 import { Component } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
@@ -12,7 +12,6 @@ import { MatDynamicFormDialogComponent } from '../../dynamic-form-dialog/dynamic
   templateUrl: './dynamic-form-icon.component.html',
   imports: [
     NgClass,
-    NgIf,
     DynamicFormColorPipe,
     DynamicFormIconPipe,
     MatDynamicFormDialogComponent,