+++ libs/material/src/lib/dynamic-form-element/dynamic-form-modal/dynamic-form-modal.component.ts
@@ -1,4 +1,3 @@
-import { NgIf } from '@angular/common';
 import { Component } from '@angular/core';
 import { DynamicFormElementComponent, DynamicFormModalBase } from '@dynamic-forms/core';
 import { MatDynamicFormDialogComponent } from '../../dynamic-form-dialog/dynamic-form-dialog.component';
@@ -7,6 +6,6 @@ import { MatDynamicFormDialogComponent } from '../../dynamic-form-dialog/dynamic
   standalone: true,
   selector: 'mat-dynamic-form-modal',
   templateUrl: './dynamic-form-modal.component.html',
-  imports: [NgIf, DynamicFormElementComponent, MatDynamicFormDialogComponent],
+  imports: [DynamicFormElementComponent, MatDynamicFormDialogComponent],
 })
 export class MatDynamicFormModalComponent extends DynamicFormModalBase {}